python/pyproject.toml
[project]
name = "stockpredictionpy"
version = "0.1.0"
description = ""
authors = [
    {name = "AbdulmalekAli60",email = "ZxshxZ123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "openpyxl (>=3.1.5,<4.0.0)",
    "curl-cffi (>=0.11.4,<0.12.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "xgboost (>=3.0.2,<4.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "stockpredictionpy", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
